SOC-3712: Only render activity of existing spaces
