+++ component/webui/src/main/java/org/exoplatform/social/webui/activity/BaseUIActivity.java
@@ -38,6 +38,7 @@ import org.exoplatform.social.core.processor.I18NActivityProcessor;
 import org.exoplatform.social.core.relationship.model.Relationship;
 import org.exoplatform.social.core.relationship.model.Relationship.Type;
 import org.exoplatform.social.core.space.SpaceException;
+import org.exoplatform.social.core.space.SpaceUtils;
 import org.exoplatform.social.core.space.model.Space;
 import org.exoplatform.social.core.space.spi.SpaceService;
 import org.exoplatform.social.core.storage.ActivityStorageException;
@@ -593,7 +594,10 @@ public class BaseUIActivity extends UIForm {
    */
   protected boolean isSpaceStreamOwner() {
     try {
-      return Utils.isSpaceActivity(this.getActivity().getId());
+      if (SpaceUtils.isSpaceActivity(this.getActivity().getId())) {
+        return SpaceUtils.isSpaceNameExisted(this.getActivity().getStreamOwner());
+      }
+      return false;
     } catch (Exception e) {
       LOG.error("Cannot get Activity ", e);
       return false;
